Add files via upload
In the project will be using 4 ultrasonic sensors, two at the front and 1 at the back. All the sensors are connected to a single Arduino Uno. All the 4 ultrasonic sensors in combination will control the car from hitting and avoiding obstacles and It will be driving on its own. There will be an algorithm written in order to choose the best move in terms of the direction where the car has to move next. In case of an obstacle on all 4 sides, the car will inform the user that there is no path to move or will wait until the path is cleared. It is completely an autonomous car which does not depend on any inputs from humans. This will further be increased to 6 sensors 2 each at the front and back and 1 each at the sides